requirements.txt
requests>=2.28.0
pandas>=1.5.0
openpyxl>=3.0.0
odfpy>=1.4.0
lxml>=4.9.0
python-dotenv>=1.0.0
entsoe-py>=0.7.0 